Fix re-render loops with JS libraries
Some JS libraries such as React mutate the dom and we can get into an
infinite re-render loop when the select is updated and the mutation
observer continually fires the reset function when it detects itself
changing. This fixes that issue by detaching the observer when resetting
the element and then re-attaching it when the reset has been completed.
This also updates the element to emit a blur event alongside the change
event when a new selection is made.